Display GET results from single datasource
The app now uses a slice as its data source. This builds up the ability to ‘add’ a new Person object. The slice acts as our data cource until persistence support is added.

NOTE ON TODO: I ould no longer use the `getAllPeople` function since I want to pass in the `people` slice around so the display happens from there instead. I couldn’t get an answer on how to do this so there’s a TODO in the code.