Visit `/add-person` to add a random person
After adding a random person, the browser is asked to go to the link `/people` so they can view the person added.

Not sure if it’s a redirect we’re looking for here. Also not sure about the status code.

In a future commit, a user can enter their own name.